requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-community>=0.0.1
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
